individual hopper and furnance implementations
